The last commit created a memory leak because 'buf' is static.  Fix that, but also make the chosen block size more consistent by basing on the larger of STDOUT vs rfd rather then just basing it on STDOUT, and change the constant 1024 to BUFSIZ.
It is unclear whether we should impose a minimum buffer size or a mod
(e.g. x % st_blksize == 0) requirement as well.  There are currently no cases
where this matters as anyone in their right mind will use 'dd' instead of
'cat' on special block devices.  But at least in regards to /proc/*/map
imposing a minimum based on fstat(rfd, ...) makes sense, though to really
be correct we should also 'fix' procfs to report a block size sufficient to
hold the map data in that case.

Reported-by: Jason Smethers <jason@smethers.net>