Correct a softupdates bug, an ir_savebp buffer was not being properly disposed of.  This bug is currently only known to effect filesystems whos block devices are not VMIO backed.  However, it's a fragile situation at best so its good to have it fixed anyhow.
ir_savebp is stashed copy of the in-memory version of an indirect file block.
It is stashed using (devvp,blkno) (whereas the actual indirect file block
is stored using (vp,-specialblkno).  For VMIO to also tickle this bug would
have required that other unrelated operations also cache (devvp,blkno), a
case which does not (yet) exist.