Make the DEBUGF() macro portable by (ugh) adding a Debug() function, which is merely printf() but to stderr.  This takes care of the caveat which lead to the use of a vararg macro -- getting everything to stderr.
Move common use of if (DEBUG(FOO)) printf... to DEBUGF(FOO, ...), using
variable length arguments to a macro.  Bump version as this makes DEBUG
statements *always* go to stderr rather than sometimes stdout.  There are
a few stragglers, which I will take care of as soon as I can.  Mostly these
relate to the need-for-death-of some of the remote job code.

Taken-From: FreeBSD/jmallett