Add a new system library, libkinfo. It is intended to replace libkvm for most userland tools, which want to deal with a living kernel, not post-mortem analysation.
Change the kern.file backing from a struct file array to a struct kinfo_file
array. The later is an independent structure, which includes the information
useful for userland. This allows changing the kernel-internals without
having to recompile libkinfo and derived programs. This is inspired by the
changes DES did in FreeBSD 5 (struct xfile).

Partly-obtained-from: FreeBSD (kern_descrip.c changes)