cond.c:1.27->1.28 make.h:1.23->1.24 parse.c:1.55->1.56 Author: harti Log: Improve make's diagnostic of mistmatched .if-.endif. This patch is slightly different from the patch in the PR. The problem is, that make handles .if clauses inside false .if clauses simply by counting them - it doesn't put them onto the conditional stack, nor even parses them so we need an extra line number stack for these ifs.
PR:             bin/61257
Submitted by:   Mikhail Teterin <mi@aldan.algebra.com>

parse.c:1.56->1.57
uthor: harti
Log:
Fix handling of comments on .elif lines. The patch given in a followup
to the PR failed, because the line skipping function is actually called
from two places in the code to do quite different things (this should
be two functions probably): in a false .if to skip to the next line
beginning with a dot and to collect .for loops. In the seconds case we
should not skip comments, because they are actually harder to handle than
we need for the .if case and should defer this to the main code.

PR:             bin/25627
Submitted by:   Seth Kingsley (original patch)

Obtained-from: FreeBSD
Submitted-by: Max Okumoto <okumoto@ucsd.edu>