Fix the boottime calculation when the time of day is set in absolute terms. We want boottime to be calculated based on the current real time minus our best uptime guess.  gd_time_seconds survives a lot of time issues and is our best uptime guess and since it is already used to calculate the new basetime, we can just assign boottime to basetime.
Note that boottime is not defined quite the same way as basetime.  The
system's internal timebase is not compensated for drift or ever jumped,
and is guarenteed to be monotonically increasing.  Drift compensation is
accomplished by adjusting basetime, so basetime will always slowly diverge
from boottime if e.g. ntpd is running.

Submitted-by: Paul Herman <pherman@frenchfries.net>