Add a intrmask_t pointer to register_int() and register_swi(), and make the interrupt thread loop set the spl for the duration of the call to the handler.
Allow interrupt threads to be run with or without a critical section based
on the kern.int_use_crit_section sysctl.  The default is to conservatively
run with a critical section for now.

Turning this off will cause cpu usage in interrupts to be properly accounted
for by top, systat, and ps.