One of the last things the system does before it tries to mount root is wait for interrupt-driven configuration hooks to finish.  This is one of the common places where the system locks up due to e.g. blown interrupt routing.
Wait up to 60 seconds for interrupt driven hooks to complete.  Generate
a warning every 10 seconds listing the hooks that have not yet completed.
Name all the hook structures to make the listing meaningful.  After 60
seconds, give up and continue booting.  The system might actually
be useable enough to aid in debugging depending on which device failed.