Revert the last device_print_child() change, it was too confusing to display the device name after the attach because recursion can insert a lot of other messages in there.
Instead, if -v is specified, generate a line before and after the attach
suitably marked [tentative] or [attached!].  This also has the advantage
of bracketing any recursion that occurs.

Keep the bootverbose printing of the device chain but get rid of the "ATTACH "
on the front.  It's obvious what it is and the attach message follows
directly now.