Journaling layer work.  Write the actual data associated with a VOP_WRITE to the journal.  Fix a number of bugs in the journal memory FIFO code.
This commit has a slight UIO hack.  Because UIOs are currently destroyed by
I/O operations on them (the iovec's get cleared), we had to make a copy of
the uio for the journaling scan.

Note that the UIO scan currently maps the data into KVA via msf_uio_iterate().
This will be made more optimal when the I/O path is fixed at some future
date to support direct page lists.