Update ACPI build wrappers to use new ACPICA code.
Submitted-by: Craig Dooley <xlnxminusx [at] gmail.com>

Add a small patch for userland ACPI tools by me, to compile
osunixxf.c . The absence of working AeLocalGetRootPointer() only
affects AcpiOsGetRootPointer() in userland ACPI tools. For kernel
module it's implemented in /sys/${ARCH}/acpica5/OsdEnvironment.c .
The callers of AcpiOsGetRootPointer() properly check anyway to see
if the returned pointer is valid, so making it no-op doesn't make
a problem.