Abstract out the routines which manipulate the mountlist.
Introduce an MP-safe mountlist scanning function.  This function keeps track
of scans which are in-progress and properly handles ripouts that occur during
the callback by advancing the matching pointers being tracked.  The callback
can safely block without confusing the scan.

This algorithm has already been successfully used for the buffer cache and
will soon be used for the vnode lists hanging off the mount point.