Stop exporting OSVERSION variable to enviornment via .makeenv. This quiets a the warning that OSVERSION is from the commandline when ports calls make recursively.
OSVERSION is only used by the make infrastructure, so anything that
references that variable would get it from make's global scope, or
would get it from a recursively called make which would have put
it into the env since it is in MAKEFLAGS.

Variables set in the command line and MAKEFLAGS are readonly, those
variables are also exported into the enviornment of make.  Nothing
in the makefiles are able to modified thier values. (POSIX)

This is compatible with .makeenv (soon to be .EXPORT:)
     o if variable does not exist create it and
       export it to the enviornment, as empty string.
     o if variable exists, export it to the enviornment
       with its value.
     o if variable that is exported is set after being
       exported, export its new value.

     o if variable is from command line or MAKEFLAGS
       do net change is value, or modify the enviornment.
       Emit a warning.